refactor: move mock-object with products and types into Lambda functions folder to view products on CloudFront
