+++ product-service/lambda-functions/product.interface.ts
